simplification of the ui + offline capable
Buttons for starting and stopping the brewing process and the pump have been merged.
SetTarget and Override Output fields merged.
API extended so that the local time of the MCU is included in every request.
Synchronization of the local browser time with the time of the MCU, so that the MCU time no longer plays a role if no online time NTP could be set.